Extract completed header task vh  from adapter
